- added showing in bucket products amount for the navigation menu

+++ src/main/java/com/geekbrains/spring/security/demo/dto/BucketDto.java
@@ -14,7 +14,7 @@ import java.util.List;
 @AllArgsConstructor
 @Builder
 public class BucketDto {
-    private double amountProducts;
+    private int amountProducts;
     private BigDecimal sum;
     private List<BucketDetailDto> bucketDetails = new ArrayList<>();
